backend/scripts/requirements.txt
# Dependências Python para o sistema de atualização diária
# Instalar com: pip install -r requirements.txt

# Manipulação de dados
pandas>=1.5.0
numpy>=1.23.0

# Banco de dados
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# HTTP requests
requests>=2.28.0

# Configuração
python-dotenv>=1.0.0

# Timezone
pytz>=2023.3

# Opcional: MySQL (se precisar suportar MySQL no futuro)
# mysql-connector-python>=8.0.0


